Add fee perc to trades csv exports
